split debug logic into DebugGuiApplication
